Move getASRouterLocalMessageIntoFromFile to fetchData.ts
